Flytt modal for dødsfalldato-registrering ut i egen fil
