+++ src/frontend/komponenter/Felleskomponenter/RegistrerDødsfallDato/RegistrerDødsfallDato.tsx
@@ -1,21 +1,9 @@
 import * as React from 'react';
 
-import styled from 'styled-components';
-
-import { Alert, Button, Modal } from '@navikt/ds-react';
 import { Dropdown } from '@navikt/ds-react';
-import type { ISODateString } from '@navikt/familie-datovelger';
-import { FamilieDatovelger } from '@navikt/familie-datovelger';
-import { FamilieTextarea } from '@navikt/familie-form-elements';
-import { RessursStatus } from '@navikt/familie-typer';
 
-import { useRegistrerDødsfallDatoSkjemaContext } from '../../../context/RegistrerDødsfallDato/RegistrerDødsfallDatoSkjemaContext';
+import RegistrerDødsfallDatoModal from './RegistrerDødsfallDatoModal';
 import type { IGrunnlagPerson } from '../../../typer/person';
-import { DatoformatNorsk } from '../../../utils/formatter';
-
-const StyledFamilieDatovelger = styled(FamilieDatovelger)`
-    margin-bottom: 1.5rem;
-`;
 
 interface IRegistrerDødsfallDato {
     person: IGrunnlagPerson;
@@ -24,26 +12,6 @@ interface IRegistrerDødsfallDato {
 
 const RegistrerDødsfallDato: React.FC<IRegistrerDødsfallDato> = ({ person, erLesevisning }) => {
     const [visModal, settVisModal] = React.useState<boolean>(false);
-
-    const {
-        skjema,
-        valideringErOk,
-        registrerManuellDødsfall,
-        nullstillSkjema,
-        settVisfeilmeldinger,
-        settRestFeil,
-        restFeil,
-    } = useRegistrerDødsfallDatoSkjemaContext({
-        onSuccess: () => settVisModal(false),
-        person,
-    });
-
-    const lukkModal = () => {
-        nullstillSkjema();
-        settVisfeilmeldinger(false);
-        settRestFeil(undefined);
-        settVisModal(false);
-    };
 
     return (
         <>
@@ -55,72 +23,11 @@ const RegistrerDødsfallDato: React.FC<IRegistrerDødsfallDato> = ({ person, erL
                 Registrer Dødsfall
             </Dropdown.Menu.List.Item>
             {visModal && (
-                <Modal
-                    open
-                    onClose={lukkModal}
-                    header={{
-                        heading: 'Registrere dødsdato',
-                        size: 'medium',
-                    }}
-                    width={'35rem'}
-                    portal
-                >
-                    <Modal.Body>
-                        <div>
-                            <StyledFamilieDatovelger
-                                {...skjema.felter.dødsfallDato?.hentNavBaseSkjemaProps(
-                                    skjema.visFeilmeldinger
-                                )}
-                                id={'registrer-døds-dato'}
-                                label={'Dødsdato'}
-                                erLesesvisning={erLesevisning}
-                                value={
-                                    skjema.felter.dødsfallDato?.verdi !== null
-                                        ? skjema.felter.dødsfallDato?.verdi
-                                        : undefined
-                                }
-                                placeholder={DatoformatNorsk.DATO}
-                                onChange={(dato?: ISODateString) =>
-                                    skjema.felter.dødsfallDato?.validerOgSettFelt(dato)
-                                }
-                            />
-                            <FamilieTextarea
-                                {...skjema.felter.begrunnelse?.hentNavBaseSkjemaProps(
-                                    skjema.visFeilmeldinger
-                                )}
-                                id={'manuell-dødsdato-begrunnelse'}
-                                label={'Begrunnelse'}
-                                erLesevisning={erLesevisning}
-                                value={skjema.felter.begrunnelse.verdi}
-                                onChange={changeEvent =>
-                                    skjema.felter.begrunnelse.validerOgSettFelt(
-                                        changeEvent.target.value
-                                    )
-                                }
-                            />
-                            {restFeil && (
-                                <Alert variant="error" style={{ marginBottom: '1.5rem' }} inline>
-                                    {restFeil}
-                                </Alert>
-                            )}
-                        </div>
-                    </Modal.Body>
-                    {!erLesevisning && (
-                        <Modal.Footer>
-                            <Button
-                                onClick={registrerManuellDødsfall}
-                                variant={valideringErOk() ? 'primary' : 'secondary'}
-                                loading={skjema.submitRessurs.status === RessursStatus.HENTER}
-                                disabled={skjema.submitRessurs.status === RessursStatus.HENTER}
-                            >
-                                Bekreft
-                            </Button>
-                            <Button onClick={lukkModal} variant={'tertiary'}>
-                                Avbryt
-                            </Button>
-                        </Modal.Footer>
-                    )}
-                </Modal>
+                <RegistrerDødsfallDatoModal
+                    lukkModal={() => settVisModal(false)}
+                    person={person}
+                    erLesevisning={erLesevisning}
+                />
             )}
         </>
     );
